requirements.txt
aiofiles==0.6.0
aiogram==2.12.1
aiohttp==3.7.4.post0
alabaster==0.7.12
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.3
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.22.0
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.6
Babel==2.9.0
beautifulsoup4==4.9.3
black==20.8b1
Brotli==1.0.9
bs4==0.0.1
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chromedriver==2.24.1
click==7.1.2
colorama==0.4.4
colored==1.4.2
colorthief==0.2.1
DateTime==4.3
decorator==4.4.2
discord==1.0.1
discord.py==1.6.0
DiscordUtils==1.2.7
discum==1.1.0
Django==3.2
dnspython==2.1.0
docutils==0.16
driver==0.1.0
emoji==1.2.0
fake-useragent==0.1.11
fanstatic==1.1
filetype==1.0.7
Flask==1.1.2
future==0.18.2
gTTS==2.2.2
idna==2.10
image==1.5.33
imageio==2.9.0
imageio-ffmpeg==0.4.3
imagesize==1.2.0
iniconfig==1.1.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
js==1.0
lazy-object-proxy==1.6.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
method==0.0.0.0.dev20200926
MouseInfo==0.1.3
moviepy==1.0.3
mtranslate==1.8
multidict==5.1.0
mypy-extensions==0.4.3
mysql-connector==2.2.9
neko==0.0.11
nekos.py==1.0.3
node==0.9.25
numpy==1.20.2
odict==1.7.0
packaging==20.9
pathspec==0.8.1
pefile==2019.4.18
Pillow==8.1.2
pipdeptree==2.0.0
pluggy==0.13.1
plumber==1.6
proglog==0.1.9
psutil==5.8.0
py==1.10.0
pyarmor==6.6.2
PyAutoGUI==0.9.52
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pyfiglet==0.8.post1
PyGetWindow==0.0.9
Pygments==2.8.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pylint==2.7.4
pymongo==3.11.3
PyMsgBox==1.0.9
PyNaCl==1.4.0
pyparsing==2.4.7
pyperclip==1.8.2
pypresence==4.2.0
pyPrivnote==0.0.1a1
PyRect==0.1.4
PyScreeze==0.1.26
pyTelegramBotAPI==3.7.7
pytest==6.2.3
python-frontmatter==0.5.0
python-obfuscator==0.0.2
PyTweening==1.0.3
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
random-user-agent==1.0.1
regex==2021.4.4
requests==2.25.1
requests-toolbelt==0.9.1
resolver==0.2.1
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
selector==0.10.1
selenium==3.141.0
shutilwhich==1.1.0
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
toml==0.10.2
tpblite==0.7.1
tqdm==4.60.0
typed-ast==1.4.3
typing-extensions==3.7.4.3
ua-parser==0.10.0
ujson==4.0.2
undetected-chromedriver==2.2.1
urllib3==1.26.2
vacefron.py==1.6.2
WebOb==1.8.7
websocket-client==0.58.0
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2021.3.3
zope.component==5.0.0
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.0.1
zope.interface==5.3.0
zope.lifecycleevent==4.3
